Project structure improved
