+++ lib/screens/share_app/share_app_screen.dart
@@ -11,8 +11,10 @@ import 'package:share/share.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:al_quran/share_icon_icons.dart';
 
-class ShareApp extends StatelessWidget {
-  const ShareApp({Key? key}) : super(key: key);
+part 'widgets/share_custom_button.dart';
+
+class ShareAppScreen extends StatelessWidget {
+  const ShareAppScreen({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -60,18 +62,18 @@ class ShareInfo extends StatelessWidget {
               textAlign: TextAlign.center,
               style: Theme.of(context).textTheme.caption),
           SizedBox(height: height * 0.05),
-          ShareCustomButton(
+          _ShareCustomButton(
             iconData: Icons.share,
             text: 'Share App',
             onPressed: () => share(context),
           ),
-          ShareCustomButton(
+          _ShareCustomButton(
             iconData: ShareIcon.github,
             text: 'GitHub Repo',
             onPressed: () =>
                 launch("https://github.com/mhmzdev/The_Holy_Quran_App"),
           ),
-          ShareCustomButton(
+          _ShareCustomButton(
             iconData: ShareIcon.googlePlay,
             text: 'Rate & Feedback',
             onPressed: () => launch(
@@ -80,54 +82,6 @@ class ShareInfo extends StatelessWidget {
           SizedBox(height: height * 0.02),
           const AppVersion()
         ],
-      ),
-    );
-  }
-}
-
-class ShareCustomButton extends StatelessWidget {
-  final String? text;
-  final IconData? iconData;
-  final void Function()? onPressed;
-  const ShareCustomButton({
-    Key? key,
-    required this.iconData,
-    required this.text,
-    required this.onPressed,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    final appProvider = Provider.of<AppProvider>(context);
-    return Padding(
-      padding: const EdgeInsets.all(8.0),
-      child: SizedBox(
-        width: MediaQuery.of(context).size.width * 0.55,
-        height: MediaQuery.of(context).size.height * 0.055,
-        child: ElevatedButton(
-          style: ButtonStyle(
-            backgroundColor: MaterialStateProperty.all(
-              AppTheme.c!.accent,
-            ),
-          ),
-          child: Row(
-            mainAxisSize: MainAxisSize.min,
-            children: <Widget>[
-              Icon(
-                iconData,
-                size: MediaQuery.of(context).size.height * 0.03,
-                color: appProvider.isDark ? Colors.black : Colors.white,
-              ),
-              Text(
-                "  $text",
-                style: TextStyle(
-                  color: appProvider.isDark ? Colors.black : Colors.white,
-                ),
-              )
-            ],
-          ),
-          onPressed: onPressed,
-        ),
       ),
     );
   }
